+++ src/test/ExSln2/LedBlinker/hal/Gpio.cs
@@ -1,4 +1,4 @@
-﻿using finlang;
+using finlang;
 
 namespace hal;