+++ src/test/ExSln2/TranspileRunner/Program.cs
@@ -1,5 +1,6 @@
 ﻿using finlang;
 using finlang.Transpiler;
+using System.Text.RegularExpressions;
 
 string thisDir = PathHelpers.GetThisDir();
 
@@ -11,6 +12,13 @@ string projectName = "LedBlinker";
 Console.WriteLine("Transpiling " + projectName + " fin/C# project...");
 
 Transpiler transpiler = new(destinationDirPath: outDir, solutionPath: slnDir + slnName, projectName: projectName);
+
+transpiler.SetFileNamer((string originalPath) =>
+{
+    originalPath = Regex.Replace(originalPath, "^mcu_avr8_", "mcu/avr8/");
+    originalPath = Regex.Replace(originalPath, "^mcu_stm32_", "mcu/stm32/");
+    return originalPath;
+});
 
 transpiler.GenerateAndWrite();
 var fileNames = transpiler.GetListOfAllGeneratedFiles();